pyproject.toml
[project]
name = "romsection"
version = "1.0.0"
license = "MIT"

requires-python = ">= 3.10"

dependencies = [
  "rtoml",
  "pyqt5",
  "numpy",
  "lru-dict",
]

[project.scripts]
romsection = "romsection.__main__:main"

[project.optional-dependencies]
dev = [
    "black",
    "pytest",
    "mypy",
    "PyQt5-stubs",
]

[tool.setuptools.package-data]
"romsection.resources" = [
    "icons/*.png",
]
